+++ src/ViewItem.js
@@ -49,6 +49,7 @@ import withLocation from './withLocation';
 import {
   itemStatusesMap,
   requestStatuses,
+  wrappingCell
 } from './constants';
 
 const requestsStatusString = map(requestStatuses, requestStatus => `"${requestStatus}"`).join(' or ');
@@ -747,10 +748,10 @@ class ViewItem extends React.Component {
                   <this.cViewMetaData metadata={item.metadata} />
                   <Row>
                     <Col xs={12}>
-                      {instance.discoverySuppress && <FormattedMessage id="ui-inventory.discoverySuppress" />}
+                      {item.discoverySuppress && <FormattedMessage id="ui-inventory.discoverySuppress" />}
                     </Col>
                   </Row>
-                  {instance.discoverySuppress && <br />}
+                  {item.discoverySuppress && <br />}
                   <Row>
                     <Col xs={2}>
                       <KeyValue
@@ -1139,9 +1140,16 @@ class ViewItem extends React.Component {
                         'Materials specified': intl.formatMessage({ id: 'ui-inventory.materialsSpecification' }),
                         'URL public note': intl.formatMessage({ id: 'ui-inventory.urlPublicNote' }),
                       }}
+                      columnWidths={{
+                        'URL relationship': '16%',
+                        'URI': '16%',
+                        'Link text': '16%',
+                        'Materials specified': '16%',
+                        'URL public note': '32%',
+                      }}
                       formatter={{
                         'URL relationship': x => refLookup(referenceTables.electronicAccessRelationships, get(x, ['relationshipId'])).name,
-                        'URI': x => <a href={get(x, ['uri'])}>{get(x, ['uri'])}</a>,
+                        'URI': x => <a href={get(x, ['uri'])} style={wrappingCell}>{get(x, ['uri'])}</a>,
                         'Link text': x => get(x, ['linkText']) || '',
                         'Materials specified': x => get(x, ['materialsSpecification']) || '',
                         'URL public note': x => get(x, ['publicNote']) || '',